Moved the PalletCard into its own component
