+++ frontend/src/pages/offers/CreatePalletModal.tsx
@@ -1,98 +1,26 @@
-import cx from 'classnames'
 import { FormEvent, FunctionComponent, useEffect, useState } from 'react'
 import Button from '../../components/Button'
 import Label from '../../components/forms/Label'
 import { CloseReason, ModalBaseProps } from '../../components/model/ModalBase'
 import PlainModal from '../../components/model/PlainModal'
+import { PALLET_CONFIGS } from '../../data/constants'
 import { PalletType } from '../../types/api-types'
+import PalletCard from './PalletCard'
 
 interface Props extends Pick<ModalBaseProps, 'onRequestClose' | 'isOpen'> {
+  /**
+   * Callback triggered by submitting the form in the modal
+   */
   onCreatePallet: (palletType: PalletType) => void
+  /**
+   * If true, the controls in the modal will stop responding to user input
+   */
+  disabled?: boolean
 }
 
-interface PalletCardProps extends PalletConfig {
-  isSelected?: boolean
-  onClick: (type: PalletType) => void
-}
-
-const PalletCard: FunctionComponent<PalletCardProps> = (props) => {
-  const cellClasses = cx('border-t-2 border-b-2 py-2 transition-colors', {
-    'border-navy-600': props.isSelected,
-    'border-transparent': !props.isSelected,
-  })
-
-  return (
-    <tr onClick={() => props.onClick(props.type)} className="cursor-pointer">
-      <td className={cx(cellClasses, 'pl-2 rounded-l border-l-2')}>
-        <label className="cursor-pointer flex items-center font-semibold">
-          <input
-            className="inline-block mr-2"
-            type="radio"
-            name="pallet-type"
-            value={props.type}
-            checked={props.isSelected}
-            onChange={() => props.onClick(props.type)}
-          />
-          {props.name}
-        </label>
-      </td>
-      <td className={cx('text-lg text-right', cellClasses)}>
-        {props.weightKg} <span className="text-gray-500">kg</span>
-      </td>
-      <td className={cx('text-lg text-right', cellClasses)}>
-        {props.lengthCm} <span className="text-gray-500">cm</span>
-      </td>
-      <td className={cx('text-lg text-right', cellClasses)}>
-        {props.widthCm} <span className="text-gray-500">cm</span>
-      </td>
-      <td
-        className={cx(
-          'text-lg text-right pr-2 rounded-r border-r-2',
-          cellClasses,
-        )}
-      >
-        {props.heightCm} <span className="text-gray-500">cm</span>
-      </td>
-    </tr>
-  )
-}
-
-interface PalletConfig {
-  name: string
-  type: PalletType
-  weightKg: number
-  lengthCm: number
-  widthCm: number
-  heightCm: number
-}
-
-const PALLET_CONFIGS: PalletConfig[] = [
-  {
-    name: 'Standard pallet',
-    type: PalletType.Standard,
-    weightKg: 700,
-    lengthCm: 120,
-    widthCm: 120,
-    heightCm: 175,
-  },
-  {
-    name: 'Euro pallet',
-    type: PalletType.Euro,
-    weightKg: 550,
-    lengthCm: 120,
-    widthCm: 80,
-    heightCm: 175,
-  },
-  {
-    name: 'Ton bag',
-    type: PalletType.Custom,
-    weightKg: 300,
-    lengthCm: 100,
-    widthCm: 100,
-    heightCm: 90,
-  },
-]
-
+/**
+ * Allow users to create a new pallet for a specific offer.
+ */
 const CreatePalletModal: FunctionComponent<Props> = (props) => {
   const [palletType, setPalletType] = useState<PalletType>(PalletType.Standard)
 
@@ -111,7 +39,9 @@ const CreatePalletModal: FunctionComponent<Props> = (props) => {
   }
 
   const selectPalletType = (type: PalletType) => {
-    setPalletType(type)
+    if (!props.disabled) {
+      setPalletType(type)
+    }
   }
 
   return (
@@ -158,10 +88,13 @@ const CreatePalletModal: FunctionComponent<Props> = (props) => {
         </table>
 
         <div className="mt-8 flex justify-end space-x-4">
-          <Button onClick={() => props.onRequestClose(CloseReason.Cancel)}>
+          <Button
+            disabled={props.disabled}
+            onClick={() => props.onRequestClose(CloseReason.Cancel)}
+          >
             Cancel
           </Button>
-          <Button type="submit" variant="primary">
+          <Button disabled={props.disabled} type="submit" variant="primary">
             Create pallet
           </Button>
         </div>
